port r50236 from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-2-10-branch@50237 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/extra/tre/regcomp.c
@@ -24,7 +24,7 @@ regncomp(regex_t *preg, const char *regex, size_t n, int cflags)
   int ret;
 #if TRE_WCHAR
   tre_char_t *wregex;
-  int wlen;
+  size_t wlen;
 
   wregex = xmalloc(sizeof(tre_char_t) * (n + 1));
   if (wregex == NULL)
@@ -90,7 +90,7 @@ regncomp(regex_t *preg, const char *regex, size_t n, int cflags)
 #endif /* TRE_MULTIBYTE */
 
   wregex[wlen] = L'\0';
-  ret = tre_compile(preg, wregex, (unsigned)wlen, cflags);
+  ret = tre_compile(preg, wregex, wlen, cflags);
   xfree(wregex);
 #else /* !TRE_WCHAR */
   ret = tre_compile(preg, (const tre_char_t *)regex, n, cflags);
@@ -110,7 +110,7 @@ tre_regcompb(regex_t *preg, const char *regex, int cflags)
 {
   int ret;
   tre_char_t *wregex;
-  int wlen, n = strlen(regex);
+  size_t wlen, n = strlen(regex);
   unsigned int i;
   const unsigned char *str = (const unsigned char *)regex;
   tre_char_t *wstr;
@@ -122,7 +122,7 @@ tre_regcompb(regex_t *preg, const char *regex, int cflags)
   for (i = 0; i < n; i++) *(wstr++) = *(str++);
   wlen = n;
   wregex[wlen] = L'\0';
-  ret = tre_compile(preg, wregex, (unsigned)wlen, cflags);
+  ret = tre_compile(preg, wregex, wlen, cflags);
   xfree(wregex);
   return ret;
 }
